+++ tools/eubfr-cli/commands/es/deleteIndices.js
