+++ tools/eubfr-cli/commands/es/showCluster.js
