+++ tools/eubfr-cli/commands/es/showDomains.js
